+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - update RegInvertIQ2 according to rx_path or tx_path inversion
 - `set_sync_word` method
 - `set_max_payload_length` method
+- `rssi_wideband` method
 
 ## [0.1.0] - 2026-03-07